refactor: 컨테이너 삭제 및 구조 재정립
컨테이너의 코드가 오히려 복잡해 보여서 컨테이너를 삭제했습니다.
Application.main() 을 의존성 주입 용으로 사용하거나, LottoController
내에서 new 키워드로 객체를 직접 생성할 예정입니다. 더 나은 방법을
찾지 못했습니다.
