+++ src/main/java/lotto/Application.java
@@ -1,16 +1,13 @@
 package lotto;
 
-import lotto.controller.LottoController;
+import lotto.lottery.controller.LottoController;
 import lotto.util.Validator;
 import lotto.view.InputHandler;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        InputHandler inputHandler = new InputHandler(new Validator());
-        LottoContainer lottoContainer = LottoContainer.getInstance();
-        LottoController lottoController = lottoContainer.lottoController();
-
+        LottoController lottoController = new LottoController();
         lottoController.run();
     }
 }
