+++ src/main/java/lotto/lottery/controller/LottoController.java
@@ -1,0 +1,7 @@
+package lotto.lottery.controller;
+
+public class LottoController {
+
+    public void run() {
+    }
+}
