+++ src/main/java/lotto/lottery/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.lottery.domain;
 
 import java.util.List;
