+++ src/main/java/lotto/lottery/domain/LottoCompany.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.lottery.domain;
 
 public class LottoCompany {
     /* TODO: 로또 회사의 역할
