+++ src/main/java/lotto/lottery/domain/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.lottery.domain;
 
 public class LottoGenerator {
 }
