+++ src/main/java/lotto/lottery/domain/Lottos.java
@@ -1,0 +1,15 @@
+package lotto.lottery.domain;
+
+import java.util.List;
+
+public class Lottos {
+    private List<Lotto> lottos;
+
+    public Lottos(List<Lotto> lottos) {
+        this.lottos = lottos;
+    }
+
+    public List<Lotto> getLottos() {
+        return lottos;
+    }
+}
